requirements.txt
scrapingbee
python-dotenv